+++ components/Sponsors/BackerWall/index.jsx
@@ -1,3 +1,5 @@
+import { useMemo } from 'react';
+
 import Avatar from '@node-core/ui-components/Common/AvatarGroup/Avatar';
 import classNames from 'classnames';
 
@@ -24,13 +26,18 @@ const shuffle = arr => [...arr].sort(() => Math.random() - 0.5);
  *   limit?: number,
  * }} props
  */
-export default ({ backers, limit = 100, showLink = true, ...props }) => (
-  <div {...props}>
-    <div className={classNames(styles.wall)}>
-      {shuffle(backers)
-        .slice(0, limit)
-        .map(backer => (
+export default ({ backers, limit = 100, showLink = true, ...props }) => {
+  const shown = useMemo(
+    () => shuffle(backers).slice(0, limit),
+    [backers, limit]
+  );
+
+  return (
+    <div {...props}>
+      <div className={classNames(styles.wall)}>
+        {shown.map(backer => (
           <Avatar
+            key={backer.slug}
             image={backer.imageUrl}
             name={backer.name}
             nickname={backer.slug}
@@ -38,16 +45,17 @@ export default ({ backers, limit = 100, showLink = true, ...props }) => (
             url={`${OC_BASE}/${backer.slug}`}
           />
         ))}
+      </div>
+      {showLink && (
+        <a
+          href={`${OC_BASE}/webpack/contributors`}
+          target="_blank"
+          rel="noreferrer noopener"
+          className={styles.link}
+        >
+          See all backers on Open Collective &rarr;
+        </a>
+      )}
     </div>
-    {showLink && (
-      <a
-        href={`${OC_BASE}/webpack/contributors`}
-        target="_blank"
-        rel="noreferrer noopener"
-        className={styles.link}
-      >
-        See all backers on Open Collective &rarr;
-      </a>
-    )}
-  </div>
-);
+  );
+};
